refactor: extract proxy HTTP transport
